src/backEnd/requirements.txt
Flask==1.0.2
requests
PyJWT
cryptography
